% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_kalman_filter_cpp}
\alias{gen_kalman_filter_cpp}
\title{Function to perform one iteration of the Kalman filter in parallel}
\usage{
gen_kalman_filter_cpp(a_0, Q_0, Q, F_, risk_sets, I_len, d, X, start, stop,
  is_event_in_bin, order_)
}
\description{
Function to perform one iteration of the Kalman filter in parallel
}

