% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.R
\name{get_cloud_means}
\alias{get_cloud_means}
\alias{get_cloud_means.PF_EM}
\alias{get_cloud_means.PF_clouds}
\title{Compute Mean Estimates from Particle Cloud}
\usage{
get_cloud_means(object, ...)

\method{get_cloud_means}{PF_EM}(object, ...)

\method{get_cloud_means}{PF_clouds}(
  object,
  cov_index = NULL,
  type = c("smoothed_clouds", "forward_clouds", "backward_clouds"),
  ...
)
}
\arguments{
\item{object}{object with class \code{PF_EM} or \code{PF_clouds}.}

\item{...}{named arguments to pass to the \code{PF_clouds} method.}

\item{cov_index}{integer vector with indices of the random effect to
include.}

\item{type}{character with the type of cloud to compute means for.}
}
\value{
A matrix which rows are time indices and columns are random effect indices.
}
\description{
Computes the estimated means from a particle cloud.
}
