% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design_matrix.R
\name{get_design_matrix}
\alias{get_design_matrix}
\title{Find the design matrix and returns it LHS and RHS}
\usage{
get_design_matrix(formula, data, response = T)
}
\description{
Find the design matrix and returns it LHS and RHS
}

