% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.R
\name{get_Q_0}
\alias{get_Q_0}
\title{Compute Time-Invariant Covariance Matrix}
\usage{
get_Q_0(Qmat, Fmat)
}
\arguments{
\item{Qmat}{covariance matrix in transition density.}

\item{Fmat}{coefficients in transition density.}
}
\value{
The invariant covariance matrix.
}
\description{
Computes the invariant covariance matrix for a vector autoregression model.
}
