% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_risk_obj_rcpp}
\alias{get_risk_obj_rcpp}
\title{Function to find the risk set in c++}
\usage{
get_risk_obj_rcpp(start, stop, event, by, start_order, max_T,
  order_by_id_and_rev_start, id)
}
\description{
Function to find the risk set in c++
}

