% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_glm.R
\name{get_survival_case_Weigths_and_data}
\alias{get_survival_case_Weigths_and_data}
\title{Function used to get design matrix and weigths for a static fit where
there are not duplicate rows but weights instead}
\usage{
get_survival_case_Weigths_and_data(formula, data, by, max_T, id, init_weights,
  risk_obj, is_for_discrete_model = T)
}
\description{
Function used to get design matrix and weigths for a static fit where
there are not duplicate rows but weights instead
}

