% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{logLik.fahrmeier_94}
\alias{logLik.fahrmeier_94}
\title{Log likelihood of smoothed state vector of a \code{fahrmeier_94} object}
\usage{
\method{logLik}{fahrmeier_94}(object, data = NULL, id, ...)
}
\arguments{
\item{object}{an object of class \code{fahrmeier_94}.}

\item{data}{new data to evaluate the likelihood for.}

\item{id}{the individual identifiers as in \code{\link{ddhazard}}.}

\item{...}{unused.}
}
\description{
Computes the log likelihood of (a potentially new) data set given the estimated:
\deqn{E_{\theta}(\alpha_1 | y_{1:d}), E_{\theta}(\alpha_{2} | y_{1:d}), ..., E_{\theta}(\alpha_{d} | y_{1:d})}

from the \code{fahrmeier_94} class object. Note that this is not the log likelihood of the observed data given the outcome.
}
