% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{logLik.PF_EM}
\alias{logLik.PF_EM}
\alias{logLik.PF_clouds}
\title{Approximate Log-Likelihood from a Particle Filter}
\usage{
\method{logLik}{PF_EM}(object, ...)

\method{logLik}{PF_clouds}(object, df = NA_real_, nobs = NA_integer_, ...)
}
\arguments{
\item{object}{an object of class \code{PF_clouds} or \code{PF_EM}.}

\item{...}{unused.}

\item{df}{degrees of freedom used in the model.}

\item{nobs}{integer with number of individuals used to estimate the
model.}
}
\value{
The approximate log-likelihood value given the observed data and set of
parameter used when simulating the clouds. An attribute
\code{"P(y_t|y_{1:(t-1)})"} has the \eqn{P(y_t|y_{1:(t-1)})} terms.
}
\description{
Computes the approximate log-likelihood using the forward filter clouds. See
the \code{vignette("Particle_filtering", "dynamichazard")} for details.
}
