% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design_matrix.R
\name{model.frame.ddformula}
\alias{model.frame.ddformula}
\alias{model.matrix.ddformula}
\title{model.frame and model.matrix for ddformula}
\usage{
\method{model.frame}{ddformula}(formula, data = NULL, subset = NULL,
  na.action = na.fail, xlev = NULL, ...)

\method{model.matrix}{ddformula}(object, data = environment(object),
  contrasts.arg = NULL, xlev = NULL, ...)
}
\arguments{
\item{formula}{Same as \code{\link{model.frame.default}}}

\item{data}{Same as \code{\link{model.matrix.default}}}

\item{subset}{Same as \code{\link{model.frame.default}}}

\item{na.action}{Same as \code{\link{model.frame.default}}}

\item{xlev}{Same as \code{\link{model.frame.default}}}

\item{...}{Unused}

\item{object}{Same as \code{\link{model.matrix.default}}}

\item{contrasts.arg}{Same as \code{\link{model.matrix.default}}}
}
\description{
Functions added to handle fixed (time-invariant) intercept coefficient for \code{\link{ddhazard}}. \code{model.frame.ddformula} always has \code{drop.unused.levels = FALSE} regardless of the input.
}
