% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ddhazard_space_errors}
\alias{plot.ddhazard_space_errors}
\title{State space error plot}
\usage{
\method{plot}{ddhazard_space_errors}(x, mod, cov_index = NA, t_index = NA,
  p_cex = par()$cex * 0.2, pch = 16, ylab = "Std. state space error",
  x_tick_loc = NA, x_tick_mark = NA, xlab = "Time", ...)
}
\arguments{
\item{x}{Result of \code{\link[=residuals.ddhazard]{residuals}} for state space errors}

\item{mod}{The \code{\link{ddhazard}} result used in the \code{\link[=residuals.ddhazard]{residuals}} call}

\item{cov_index}{The indices of state vector errors to plot. Default is to use all which is likely what you want if the state space errors are standardized}

\item{t_index}{The bin indices to plot. Default is to use all bins}

\item{p_cex}{\code{cex} argument for the points}

\item{pch, ylab, xlab}{Arguments to override defaults set in the function}

\item{x_tick_loc, x_tick_mark}{\code{at} and \code{labels} arguments passed to \code{axis}}

\item{...}{Arguments passed to plot}
}
\description{
Plot function for state space errors from \code{\link{ddhazard}} fit
}
