% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fahrmeier_94_stdSpaceErrors}
\alias{plot.fahrmeier_94_stdSpaceErrors}
\title{Plot function for standadized statespace errors from ddhazardfit}
\usage{
\method{plot}{fahrmeier_94_stdSpaceErrors}(object, mod, vars_ = NA,
  t_index = NA, p_cex = par()$cex * 0.2, pch = 16,
  ylab = "Std. state space error", x_tick_loc = NA, x_tick_mark = NA,
  xlab = "Time", ...)
}
\description{
Plot function for standadized statespace errors from ddhazardfit
}

