% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PF_clouds}
\alias{plot.PF_clouds}
\title{Plot of clouds from a \code{PF_clouds} object}
\usage{
\method{plot}{PF_clouds}(x, y, type = c("smoothed_clouds", "forward_clouds",
  "backward_clouds"), ylim, add = FALSE, qlvls = c(0.05, 0.5, 0.95),
  pch = 4, lty = 1, ..., cov_index)
}
\arguments{
\item{x}{an object of class \code{PF_clouds}.}

\item{y}{unused.}

\item{type}{parameter to specify which cloud to plot.}

\item{ylim}{\code{ylim} passed to \code{\link{matplot}}.}

\item{add}{\code{TRUE} if a new plot should not be made.}

\item{qlvls}{vector of quantile levels to be plotted.}

\item{pch}{\code{pch} argument for the quantile points.}

\item{lty}{\code{lty} argument for the mean curves.}

\item{...}{unused.}

\item{cov_index}{indices of the state vector to plot. All are plotted if this argument is omitted.}
}
\value{
List with quantile levels and mean curve.
}
\description{
Plots mean curve along with quantiles through time for the forward, backward or smoothed clouds.
}
