% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PF_EM}
\alias{plot.PF_EM}
\title{Plot for a PF_EM Object}
\usage{
\method{plot}{PF_EM}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{PF_EM}.}

\item{y}{unused.}

\item{...}{arguments to \code{\link{plot.PF_clouds}}.}
}
\value{
See \code{\link{plot.PF_clouds}}
}
\description{
Short hand to call \code{\link{plot.PF_clouds}}.
}
