% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fahrmeier_94}
\alias{predict.fahrmeier_94}
\title{Predict function for result of ddhazard}
\usage{
\method{predict}{fahrmeier_94}(object, new_data, type = c("response", "term"),
  tstart = "start", tstop = "stop", use_parallel = F, sds = F, ...)
}
\description{
Predict function for result of ddhazard
}

