% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ddhazard}
\alias{print.ddhazard}
\title{Print function for \code{ddhazard} result}
\usage{
\method{print}{ddhazard}(x)
}
\arguments{
\item{x}{object returned from \code{\link{ddhazard}}.}

\item{var_indices}{variable indices to print for time-varying effects.}

\item{time_indices}{time intervals to print for time-varying effects.}

\item{digits}{number of digits to print.}

\item{...}{not used.}
}
\description{
The \code{sd} printed for time-varying effects are point-wise standard deviations from the smoothed covariance matrices.
}
