% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ddhazard_boot}
\alias{print.ddhazard_boot}
\title{Summary Statistics for a ddhazard_boot Object}
\usage{
\method{print}{ddhazard_boot}(x, digits = getOption("digits"), index = 1L:ncol(boot.out$t), ...)
}
\arguments{
\item{x}{returned object from a \code{\link{ddhazard_boot}} call.}

\item{digits}{the number of digits to be printed in the summary statistics.}

\item{index}{indices indicating for which elements of the bootstrap output summary statistics are required.}

\item{...}{not used.}
}
\description{
Arguments have the same effects as for an object from a \code{\link{boot}} call. See \code{\link[=print.boot]{print}}.
}
\seealso{
\code{\link{ddhazard_boot}}
}
