% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.fahrmeier_94}
\alias{residuals.fahrmeier_94}
\title{Residuals function for result of ddhazard fit}
\usage{
\method{residuals}{fahrmeier_94}(object, type = c("stdSpaceErrors",
  "Pearson"), data_)
}
\description{
Residuals function for result of ddhazard fit
}

