% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design_matrix.R
\name{ddFixed}
\alias{ddFixed}
\title{Function used in formula of \code{\link{ddhazard}} for time-invariant effects}
\usage{
ddFixed(object)
}
\arguments{
\item{object}{Expression that would be used in formula. E.g. \code{x} or \code{poly(x, degree = 3)}}
}
\description{
Function used in formula of \code{\link{ddhazard}} for time-invariant effects
}

