% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddhazard.R
\name{ddhazard}
\alias{ddhazard}
\title{Function to fit dynamic discrete hazard model}
\usage{
ddhazard(formula, data, by, a_0, Q_0, id, Q = Q_0, n_max = 10^2,
  eps = 10^-3, save_all_output = F, max_T, save_risk_set = F,
  order_ = 1, est_Q_0 = T, verbose = F, method = "EKF",
  model = "logit")
}
\description{
Function to fit dynamic discrete hazard model
}

