% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddhazard.R
\name{ddhazard}
\alias{ddhazard}
\title{Function to fit dynamic discrete hazard models}
\usage{
ddhazard(formula, data, model = "logit", by, max_T, id, a_0, Q_0, Q = Q_0,
  order = 1, control = list(), verbose = F)
}
\arguments{
\item{formula}{\code{\link[survival]{coxph}} like formula with \code{\link[survival]{Surv}(tstart, tstop, event)} on the left hand site of \code{~}}

\item{data}{Data frame or environment containing the outcome and co-variates}

\item{model}{\code{"logit"}, \code{"exponential_trunc_time_only"}, \code{"exponential_binary_only"} or \code{"exp_combined"} for the discrete time function using the logistic link function in the first case or for the continuous time model with different estimation method in the three latter cases (see the ddhazard for details on the methods)}

\item{by}{Interval length of the bins in which parameters are fixed}

\item{max_T}{End of the last interval. The last stop time with an event is selected if the parameter is omitted}

\item{id}{Vector of ids for each row of the in the design matrix}

\item{a_0}{Vector \eqn{a_0} for the initial coefficient vector for the first iteration (optional). Default is estimates from static model (see \code{\link{static_glm}})}

\item{Q_0}{Covariance matrix for the prior distribution}

\item{Q}{Initial covariance matrix for the state equation}

\item{order}{Order of the random walk}

\item{control}{List of control variables (see details below)}

\item{verbose}{\code{TRUE} if you want status messages during execution}
}
\value{
A list with class \code{fahrmeier_94}. The list contains:
\tabular{ll}{
\code{formula}\verb{ }\tab The passed formula \cr
\code{state_vecs}\verb{ }\tab 2D matrix with the estimated state vectors (regression parameters) in each bin \cr
\code{state_vars}\verb{ }\tab 3D array with smoothed variance estimates for each state vector \cr
\code{lag_one_cov}\verb{ }\tab 3D array with lagged correlation matrix for each for each change in the state vector. Only present when the model is logit and the method is EKF \cr
\code{n_risk}\verb{ }\tab The number of observations in each interval \cr
\code{times}\verb{ }\tab The interval borders \cr
\code{risk_set}\verb{ }\tab The object from \code{\link{get_risk_obj}} if saved \cr
\code{data}\verb{ }\tab The \code{data} argument if saved \cr
\code{order}\verb{ }\tab Order of the random walk \cr
\code{F_}\verb{ }\tab Matrix with that map transition from one state vector to the next \cr
\code{method}\verb{ }\tab Method used in the E-step \cr
\code{est_Q_0}\verb{ }\tab\code{TRUE} if \code{Q_0} was estimated in the EM-algorithm \cr
\code{hazard_func}\verb{ }\tab Hazard function \cr
\code{hazard_first_deriv}\verb{ }\tab First derivative of the hazard function with respect to the linear predictor
}
}
\description{
Function to fit dynamic discrete hazard models using state space models
}
\details{
This function can be used to estimate a binary regression where the regression parameters follows a given order random walk. The order is specified by the \code{order} argument. 1. and 2. order random walks is implemented. The regression parameters are updated at time \code{by}, 2\code{by}, ..., \code{max_T}. See the vignette 'ddhazard' for more details

The Extended Kalman filter or Unscented Kalman filter needs an initial co-variance matrix \code{Q_0} and state vector \code{a_0}. An estimate from a time-invariant model is provided for \code{a_0} if it is not supplied (the same model you would get from \code{\link{static_glm}} function). A diagonal matrix with large entries is recommended for \code{Q_0}. What is large dependents on the data set and \code{model}. Further, a variance matrix for the first iteration \code{Q} is needed. It is recommended to select diagonal matrix with low values for the latter. The \code{Q}, \code{a_0} and optionally \code{Q_0} is estimated with an EM-algorithm

The model is specified through the \code{model} argument. Currently, \code{'logit'} and \code{'exponential'} is available. The former uses an logistic model where outcomes are binned into the intervals. Be aware that there can be loss of information due to binning. It is key for the logit model that the \code{id} argument is provided if individuals in the data set have time varying co-variates. The latter model uses an exponential model for the arrival times where there is no loss information due to binning

It is recommended to see the Shiny app demo for this function by calling \code{\link{ddhazard_app}()}
}
\section{Control}{

The \code{control} argument allows you to pass a \code{list} to select additional parameters. See the vignette 'ddhazard' for more information on hyper parameters. Unspecified elements of the list will yield default values
\tabular{ll}{
\code{method}\verb{  }\tab Set to the method to use in the E-step. Either \code{"EKF"} for the Extended Kalman Filter or \code{"UKF"}for the Unscented Kalman Filter. \code{"EKF"} is the default \cr
\code{LR}\verb{  }\tab Learning rate for the Extended Kalman filter \cr
\code{NR_eps}\verb{  }\tab Tolerance for the Extended Kalman filter. Default is \code{NULL} which means that no extra iteration is made in the correction step \cr
\code{alpha}\verb{  }\tab Hyper parameter \eqn{\alpha} in the Unscented Kalman Filter \cr
\code{beta}\verb{  }\tab Hyper parameter \eqn{\beta} in the Unscented Kalman Filter  \cr
\code{kappa}\verb{  }\tab Hyper parameter \eqn{\kappa} in the Unscented Kalman Filter \cr
\code{n_max}\verb{  }\tab Maximum number of iteration in the EM-algorithm \cr
\code{eps}\verb{  }\tab Tolerance parameter for the EM-algorithm\cr
\code{est_Q_0}\verb{  }\tab\code{TRUE} if you want the EM-algorithm to estimate \code{Q_0}. Default is \code{FALSE}\cr
\code{save_risk_set}\verb{  }\tab \code{TRUE} if you want to save the list from \code{\link{get_risk_obj}} used to estimate the model. It may be needed for later call to \code{residuals}, \code{plot} and \code{logLike}. Can be set to \code{FALSE} to save memory\cr
\code{save_data}\verb{  }\tab \code{TRUE} if you want to save the list \code{data} argument. It may be needed for later call to \code{residuals}, \code{plot} and \code{logLike}. Can be set to \code{FALSE} to save memory \cr
\code{ridge_eps}\verb{ }\tab Penalty term added to the diagonal of the covariance matrix of the observational equation in either the EKF or UKF \cr
\code{fixed_terms_method}\verb{ }\tab The method used to estimate the fixed effects. Either \code{'M_step'} or \code{'E_step'} for estimation in the M-step or E-step respectively \cr
\code{Q_0_term_for_fixed_E_step}\verb{ }\tab The diagonal value of the initial covariance matrix, \code{Q_0}, for the fixed effects if fixed effects are estimated in the E-step \cr
\code{eps_fixed_parems}\verb{ }\tab Tolerance used in the M-step of the Fisher's Scoring Algorithm for the fixed effects
}
}
\references{
Fahrmeir, Ludwig. \emph{Dynamic modelling and penalized likelihood estimation for discrete time survival data}. Biometrika 81.2 (1994): 317-330.

Durbin, James, and Siem Jan Koopman. \emph{Time series analysis by state space methods}. No. 38. Oxford University Press, 2012.
}
\seealso{
\code{\link[=plot.fahrmeier_94]{plot}}, \code{\link[=residuals.fahrmeier_94]{residuals}}, \code{\link[=predict.fahrmeier_94]{predict}}, \code{\link{static_glm}}, \code{\link{ddhazard_app}}
}

