% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{ddhazard_app}
\alias{ddhazard_app}
\title{A Shiny app to illustrates model and method}
\usage{
ddhazard_app()
}
\description{
Runs a shiny app where you try different model specification on simulated data
}
\examples{
\dontrun{
dynamichazard::ddhazard_app
}
}

