% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.R
\name{get_Q_0}
\alias{get_Q_0}
\title{Compute Invariant Covaraince Matrix}
\usage{
get_Q_0(Qmat, Fmat)
}
\arguments{
\item{Qmat}{Covaraince matrix in transition density.}

\item{Fmat}{Coefficients in transition density.}
}
\value{
The invariant covaraince matrix.
}
\description{
Computes the invariant covaraince matrix for a vector autoregression model.
}
