% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{logLik.PF_clouds}
\alias{logLik.PF_clouds}
\title{Log Likelihood of a PF_clouds Object}
\usage{
\method{logLik}{PF_clouds}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PF_clouds}.}

\item{...}{unused.}
}
\value{
The log-likelihood value given the observed data and set of parameter used when simulating the clouds.
}
\description{
Computes the log likelihood using the forward filter clouds. See the \code{vignette("Particle_filtering", "dynamichazard")} for details.
}
