% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.R
\name{PF_EM}
\alias{PF_EM}
\title{EM estimation with particle filters}
\usage{
PF_EM(formula, data, model = "logit", by, max_T, id, a_0, Q_0, Q, order = 1,
  control = list(), trace = 0)
}
\arguments{
\item{formula}{\code{\link[survival]{coxph}} like formula with \code{\link[survival]{Surv}(tstart, tstop, event)} on the left hand site of \code{~}}

\item{data}{Data frame or environment containing the outcome and co-variates}

\item{model}{\code{"logit"}, \code{"exp_clip_time_w_jump"}, \code{"exp_clip_time"} or \code{"exp_bin"} for the discrete time function using the logistic link function in the first case or for the continuous time model with different estimation method in the three latter cases (see the ddhazard vignette for details of the methods)}

\item{by}{Interval length of the bins in which parameters are fixed}

\item{max_T}{End of the last interval. The last stop time with an event is selected if the parameter is omitted}

\item{id}{Vector of ids for each row of the in the design matrix}

\item{a_0}{Vector \eqn{a_0} for the initial coefficient vector for the first iteration (optional). Default is estimates from static model (see \code{\link{static_glm}})}

\item{Q_0}{Covariance matrix for the prior distribution}

\item{Q}{Initial covariance matrix for the state equation}

\item{order}{Order of the random walk}

\item{control}{List of control variables (see the control section below)}

\item{trace}{Argument to get progress information. Zero will yield no info an larger integer values will yield incrementally more information.}
}
\value{
TODO: write me
}
\description{
Method to estimate the hyper parameters with an EM algorithm.
}
\details{
See the particle_filter vignette for details.
}
\section{Control}{

The \code{control} argument allows you to pass a \code{list} to select additional parameters. See the vignette 'ddhazard' for more information on hyper parameters. Unspecified elements of the list will yield default values
\describe{
\item{\code{method}}{Method for forward, backward and smoothing filter. See the particle_filter vignette for details.}
\item{\code{N_fw_n_bw}}{Number of particles to use in forward and backward filter.}
\item{\code{N_first}}{Number of particles to use at time \eqn{0} and time \eqn{d + 1}.}
\item{\code{N_smooth}}{Number of particles to use in particle smoother.}
\item{\code{eps}}{Convergence threhshold in EM method.}
\item{\code{n_max}}{Maximum number of iterations of the EM algorithm.}
\item{\code{n_threads}}{Maximum number threads to use in the computations.}
\item{\code{forward_backward_ESS_threshold}}{Required effective sample size to not re-sample in the particle filters.}
}
}

