% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fahrmeier_94}
\alias{plot.fahrmeier_94}
\title{Plots to illustrate the estimate state space variables from a ddhazardfit}
\usage{
\method{plot}{fahrmeier_94}(object, new_data, xlab = "Time",
  ylab = "Hazard", type = "cov", plot_type = "l", cov_index, ylim,
  col = "black", add = F, ...)
}
\description{
Plots to illustrate the estimate state space variables from a ddhazardfit
}

