% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fahrmeier_94}
\alias{plot.fahrmeier_94}
\title{Plots for \code{\link{ddhazard}}}
\usage{
\method{plot}{fahrmeier_94}(x, xlab = "Time", ylab = "Hazard",
  type = "cov", plot_type = "l", cov_index, ylim, col = "black",
  add = F, do_alter_mfcol = T, ...)
}
\arguments{
\item{x}{Result of \code{\link{ddhazard}} call}

\item{xlab, ylab, ylim, col}{Arguments to overide defaults set in the function}

\item{type}{Type of plot. Currently, only \code{"cov"} is available for plot of the state space parameters}

\item{plot_type}{The \code{type} argument passed to \code{plot}}

\item{cov_index}{The index (indices) of the state space parameter(s) to plot}

\item{add}{\code{FALSE} if you want to make a new plot}

\item{do_alter_mfcol}{\code{TRUE} if the function should alter \code{par(mfcol)} in case that \code{cov_index} has more than one element}

\item{...}{Arguments passed to \code{plot} or \code{lines} depending on the value of \code{add}}
}
\description{
Plot to illustrate the estimate state space variables from a \code{\link{ddhazard}} fit
}
\details{
Creates a plot of state variables or adds state variables to a plot with indices \code{cov_index}. Pointwise 1.96 std. confidence intervals are provided with the smoothed co-variance matrices from the fit
}

