% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ddhazard}
\alias{print.ddhazard}
\title{Print function for \code{ddhazard} result}
\usage{
\method{print}{ddhazard}(x, var_indices = 1:ncol(x$state_vecs),
  time_indices = 1:nrow(x$state_vecs), digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object returned from \code{\link{ddhazard}}}

\item{var_indices}{Variable indices to print for time-varying effects}

\item{time_indices}{Time intervals to print for time-varying effects}

\item{digits}{Number of digits to print}

\item{...}{Not used}
}
\description{
The \code{sd} printed for time-varying effects are point-wise standard deviations from either the filter with smoothing
}
