% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_glm.R
\name{static_glm}
\alias{static_glm}
\title{Function to make a static glm fit from a input with a \code{Surv} object
as the right hand site of forumla}
\usage{
static_glm(formula, data, by, max_T, id, family = "binomial", model = F,
  weights, risk_obj, ...)
}
\description{
Function to make a static glm fit from a input with a \code{Surv} object
as the right hand site of forumla
}

